Add pkeyauthChallengeFactory and refactor the related callings to pkeyauth challenge.
